backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
supabase==2.5.0
pydantic==2.7.1
python-dotenv==1.0.1
python-jose==3.3.0
langchain==0.2.1
langchain-openai==0.1.7
langchain-community==0.2.1
requests
gymnasium
stable-baselines3
psutil
ultralytics
opencv-python
python-multipart
reportlab
pmdarima
pandas
